refactor: replace feedback slide-out panel with dedicated /feedback page
The sidebar feedback button opened a slide-out panel, which felt
inconsistent with every other sidebar item (all navigate to pages).
Replace it with a dedicated /feedback page and a standard NavLink.

- New /feedback page with inline Canny widget
- Sidebar: replace button + panel state with a NavLink
- Delete feedback-widget.tsx (slide-out panel)
- Update i18n keys, smoke tests, a11y scan
- Add CANNY vars to .env.example
- Update changelog to reflect new UX

Fixes #149
