+++ .env.example
@@ -37,6 +37,12 @@ TURSO_AUTH_TOKEN=
 # Free tier: 15 req/min, 1M tokens/day
 GOOGLE_GENERATIVE_AI_API_KEY=
 
+# Canny Feedback Widget — https://developers.canny.io
+# Board token (public): find at https://developers.canny.io/install/widget/web
+# SSO private key (server-only): find at Settings → Single Sign-On in your Canny dashboard
+NEXT_PUBLIC_CANNY_BOARD_TOKEN=
+CANNY_SSO_PRIVATE_KEY=
+
 # ─── Demo / Preview Mode ─────────────────────────────────────────────────────
 # Set to "true" to enable demo mode:
 #   - Fake auth (no Discord OAuth required)
