+++ changelog/en/2026-04-04-canny-feedback-widget.mdx
@@ -5,20 +5,7 @@ title: "Send feedback directly from the app"
 tags: ["feature"]
 ---
 
-There's now a **Feedback** button in the sidebar. Click it to open a slide-out panel where you can submit feature requests, report bugs, or vote on ideas from other players — all without leaving the app.
+There's now a **Feedback** page in the app. Click "Feedback" in the sidebar to submit feature requests, report bugs, or vote on ideas from other players.
 
-- **Sidebar feedback button**: A new "Feedback" item sits in the sidebar between Settings and Admin. One click opens the panel; press Escape or click outside to close it.
-- **Integrated feedback board**: The panel shows the full feedback board where you can post suggestions, upvote existing ones, and track their status.
+- **Dedicated feedback page**: A new "Feedback" link in the sidebar takes you to a full page where you can browse and post suggestions, upvote existing ones, and track their status.
 - **Automatically linked to your account**: Your feedback is tied to your account, so you'll get notified when something you requested ships.
-
-<div className="changelog-image-grid">
-
-![Before: dashboard with no feedback option](/changelog/canny-feedback-widget/before-dashboard.png)
-
-![After: new Feedback button in the sidebar](/changelog/canny-feedback-widget/after-sidebar-annotated.png)
-
-</div>
-
-### Feedback panel
-
-![Feedback panel slides out from the right with the full board](/changelog/canny-feedback-widget/after-panel-annotated.png)
