+++ changelog/es/2026-04-04-canny-feedback-widget.mdx
@@ -5,20 +5,7 @@ title: "Envía sugerencias directamente desde la app"
 tags: ["feature"]
 ---
 
-Ahora hay un botón de **Sugerencias** en la barra lateral. Haz clic para abrir un panel deslizante donde puedes enviar solicitudes de funciones, reportar errores o votar las ideas de otros jugadores — todo sin salir de la app.
+Ahora hay una página de **Sugerencias** en la app. Haz clic en "Sugerencias" en la barra lateral para enviar solicitudes de funciones, reportar errores o votar las ideas de otros jugadores.
 
-- **Botón de sugerencias en la barra lateral**: Un nuevo elemento "Sugerencias" se encuentra en la barra lateral entre Ajustes y Admin. Un clic abre el panel; pulsa Escape o haz clic fuera para cerrarlo.
-- **Tablero de sugerencias integrado**: El panel muestra el tablero completo donde puedes publicar sugerencias, votar las existentes y seguir su estado.
+- **Página dedicada de sugerencias**: Un nuevo enlace "Sugerencias" en la barra lateral te lleva a una página completa donde puedes ver y publicar sugerencias, votar las existentes y seguir su estado.
 - **Vinculado automáticamente a tu cuenta**: Tus sugerencias están asociadas a tu cuenta, así que recibirás notificaciones cuando algo que pediste se implemente.
-
-<div className="changelog-image-grid">
-
-![Antes: panel sin opción de sugerencias](/changelog/canny-feedback-widget/before-dashboard.png)
-
-![Después: nuevo botón de sugerencias en la barra lateral](/changelog/canny-feedback-widget/after-sidebar-annotated.png)
-
-</div>
-
-### Panel de sugerencias
-
-![El panel de sugerencias se desliza desde la derecha con el tablero completo](/changelog/canny-feedback-widget/after-panel-annotated.png)
